Add Track and Set Student Progress
Added a new track and set student progress feature that allows academic officers to set recovering students grades and course attempt numbers while they're re-attempting their failed courses

+++ App/src/main/java/system/app/Student.java
@@ -86,7 +86,7 @@ public class Student {
         return failedCourses;
     }
     
-    public double getFailedCoursesCount() {
+    public int getFailedCoursesCount() {
         return failedCourses.size();
     }
     public List<CourseEnrollment> getEnrollment() {
